+++ GeoSpatialTools/__init__.py
@@ -3,7 +3,9 @@
 from .neighbours import find_nearest
 from .distance_metrics import haversine
 from .great_circle import GreatCircle
-from .quadtree import Ellipse, QuadTree, Record, Rectangle
+from .record import Record, SpaceTimeRecord
+from .quadtree import QuadTree
+from .shape import Ellipse, Rectangle, SpaceTimeEllipse, SpaceTimeRectangle
 from .kdtree import KDTree
 
 __all__ = [
@@ -13,6 +15,9 @@ __all__ = [
     "QuadTree",
     "Record",
     "Rectangle",
+    "SpaceTimeEllipse",
+    "SpaceTimeRecord",
+    "SpaceTimeRectangle",
     "find_nearest",
     "haversine",
 ]